paragon id Int to Long change

+++ app/src/main/java/com/example/photoapp/features/paragon/data/ProduktParagonDao.kt
@@ -13,7 +13,7 @@ interface ProduktParagonDao {
     fun getAll(): List<ProduktParagon>
 
     @Query("SELECT * FROM ProduktParagon WHERE paragonId = :paragonId")
-    fun getByParagonId(paragonId: Int): List<ProduktParagon>
+    fun getByParagonId(paragonId: Long): List<ProduktParagon>
 
     @Insert
     fun insert(produkt: ProduktParagon): Long
